+++ test/flutter_local_ai_platform_interface_test.dart
@@ -43,6 +43,11 @@ class ImplementsFlutterLocalAiPlatform implements FlutterLocalAiPlatform {
   }
 
   @override
+  Future<String> availabilityReason() {
+    throw UnimplementedError();
+  }
+
+  @override
   Future<bool> openAICorePlayStore() {
     throw UnimplementedError();
   }